+++ changes.md
@@ -11,7 +11,7 @@ This patch fixes a number of issues related to UI.
     * Power tips (the descriptions of any applied powers in the buffer) will no longer incorrectly show formatting information if that information occurs at the start of the first word of the effect description (for example #y at the beginning of weak's description)
     * Monsters now have location information
 * Potions
-    * Potion location is now properly announced (excluding compendium)
+    * Potion location is now properly announced
     * Fixed a bug where you could not read the tooltip for an empty potion slot
     * Potion rarity has been added to the potion buffer
 * Relics